+++ src/shared/ui/Input/index.tsx
@@ -2,7 +2,6 @@ import React, { HTMLInputTypeAttribute } from "react"
 
 import classNames from "classnames"
 import { Form } from "react-bootstrap"
-import { Icon } from "react-bootstrap-icons"
 import { useController, useFormContext } from "react-hook-form"
 
 import "./style.scss"
@@ -16,7 +15,7 @@ interface IInputCommonProps {
   onChange?: (e: string | number) => void
   className?: string
 
-  Icon?: Icon
+  Icon?: JSX.Element
   onIconClick?: () => void
 }
 
@@ -52,7 +51,7 @@ const Input = (props: IInputCommonProps) => {
 
       {Icon && onIconClick && (
         <div className={classNames("input-icon")} onClick={onIconClick}>
-          <Icon />
+          {Icon}
         </div>
       )}
