+++ src/shared/ui/PasswordIcon/index.tsx
@@ -1,0 +1,9 @@
+import { EyeSlash, Eye } from "react-bootstrap-icons"
+
+interface PasswordIconProps {
+  isSecure: boolean
+}
+
+export const PasswordIcon = ({ isSecure }: PasswordIconProps) => {
+  return isSecure ? <EyeSlash /> : <Eye />
+}
