+++ src/shared/ui/index.ts
@@ -9,3 +9,4 @@ export { default as NavbarButton } from "./NavbarButton"
 export { default as Avatar } from "./Avatar"
 export { default as CustomCard } from "./Card"
 export { DisplaySystemMessage } from "./DisplaySystemMessage"
+export { PasswordIcon } from "./PasswordIcon"
