+++ src/shared/utils/hooks/usePasswordType.ts
@@ -1,0 +1,17 @@
+import { useState } from "react"
+
+type PasswordInputType = "text" | "password"
+
+export const usePasswordType = (): [PasswordInputType, () => void] => {
+  const [isPasswordType, setIsPasswordType] = useState<PasswordInputType>("password")
+
+  const onPasswordIconClick = () => {
+    if (isPasswordType === "password") {
+      setIsPasswordType("text")
+    } else {
+      setIsPasswordType("password")
+    }
+  }
+
+  return [isPasswordType, onPasswordIconClick]
+}
